integrated interrupt controller
Added ~IRQ input to core module, and changed the
wiring of the CPU's ~IRQ input to feed in the ~IRQ
output of the interrupt controller.  (A pull-up
resistor will keep this input high when the interrupt
controller isn't installed.)

Have the peripheral module send the Timer 1 output
to the interrupt controller's ~IRQ3 input.  This will
be the system periodic timer interrupt.

Also have the peripheral module export the pins of
Port C, and use PC0 and PC1 to clear the flip flops
that latch a pulse on ~IRQ3 and ~IRQ6, which are
edge-triggered.