generate pixels from buffer in block RAM
Reading from the block RAM seems to work fine.
However, when the command code tries to write to the
block RAM, icestorm fails to infer that the memory
should be a block RAM. (Actually, I'm not 100%
confident that it's inferring this correctly in the
case where there are only reads from the buffer.)

More investigation needed.