HOLY CRAP WE ARE DISPLAYING PIXEL DATA
Data from the command FIFO is successfully copied into the
line buffer BRAM, and in turn the contents of the line buffer
BRAM are used to control generation of foreground/background
pixel colors.

Even better, the design could even work at an internal clock
frequency of 40 MHz, because loading data from the BRAM and
advancing the read address and pixel state is done ahead of
time. (The BRAM is capable of accepting and producing a
data value on each clock.)