working on command loop, has some bugs
For now, we're just trying to read data from the FIFO
and interpret as a command byte followed by one or more
data bytes. The state machine doesn't seem to work correctly,
because the transition from recognizing a valid command byte
to processing its associated data byte doesn't seem to be
handled correctly.

Added a couple of debug LEDs to the Upduino to make it easier
to see what is going on.