more futzing with sync generation
Also, I made the pixels where hcount is a multiple of 8
magenta, and interestingly, these show up in the expected
place. So, it just seems like vcount is being weird, where
the first visible line has vcount of 1 rather than 0.
I think the thing to do might be to just have the syncgen2
module "lie" about vcount. That will be the next thing
to try.