compile ch_col and ch_row on the fly from hcount/vcount
This gives us output that is very, very close to correct!

There is still a minor issue with the upper-left character,
I think because vcount is not 0 when the 8 pixels of the top
row of that character are generated. This should be pretty
easy to work around.

Another issue is that a half-row of characters is displayed
on the bottom 8 rows of pixels. That should also be fairly
easy to deal with.