attempting to rewrite command processing
Goal is to cleanly split the handling of command vs.
data bytes, vs. running a state machine in response to
a received command.

The idea is that there is now a separate state machine for
each command type, and the main data reception loop sends
bytes via a shared_reg.

Very much not working at the moment, and it doesn't
even seem to properly separate command vs. data bytes.