lots of changes to TMS9918A and YM2149 modules
Changed the TMS9918A circuit to use -RMEM and -WMEM rather than R/-W,
since -RMEM and -WMEM are properly synchronized with E and the overall
CPU bus cycle. (Duh!)

Also chnaged the YM2149 circuit to use -RMEM and -WMEM. However,
it's less clear that this is a desirable change, because we would
like BDIR and BC1 to change "at the same time", and only when
the chip is being accessed. However, if the chip select is asserted
before -RMEM/-WMEM are asserted, then meaningless BDIR and BC1
values are going to be generated, which is not good. Need to think
about this.