got the VRAM test working
Turns out "cmpa B" does *not* compare the contents of A
and B. Instead, it compares A to (I think) the byte of
the direct memory page indexed by B.

Fixed the problem by comparing the byte read from VRAM
into A with the contents of a memory variable containing
the expected value.

We can now be reasonably confident that the host system
can access VRAM, both for writing and reading.