bugfix: ReadoutA0 wasn't connected
A wire was misplaced on the main sheet, which meant
that the ReadoutA0 input of the VRAM module had no
connection to the ReadoutA0 output from the Readout
module.

This will require a bodge wire on the 29-Jul-2024
version of the PCB, but fortunately it will be a very
easy one, since the 74ALS163 that generates ReadoutA0
is on the bottom of the PCB and can be easily connected
to the appropriate pin on either VRAM PLCC socket.